Implement `gem fetch` with dependency installation...
The `gem fetch` command will fetch a gem and save it to disk, but it doesn't
follow dependencies.  Having to hand-maintain the full dependency list for
every gem we included was currently cheap, but isn't likely to stay that
way forever.

This is a trivial implementation of `gem fetch` that also follows
dependencies, so that we can stop worrying about that problem.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>