Use p7zip, if available, to extract the ISO image
The p7zip tool - which is widely available - can extract an ISO image like any
other container format.  This changes the script to detect it and, if
available, prefer it to a loopback mount to extract the content of the
ISO image.

This makes it easier to rebuild the microkernel without root privileges,
something that will make it much easier to have the build process driven out
of our CI system.

The older behaviour is still retained, and new error handling is introduced to
detect a situation where neither root or p7zip are available, and to report a
clear error to the user.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>