Derive the ISO image version from git metadata
This updates the bundle builder, and the ISO rebuild scripts, to use a version
number derived from the git repository.  It uses the tags to assign the major
version, and supports nightly as well as full builds.

This helps eliminate the human input requirement to rebuilding the ISO image,
which makes it easier to get the process automated via CI.

Manually specifying the version still works, and will override the version
number derived from git.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>