Avoid chroot when running the guest busybox
Using the guest busybox is nice, but it is easier if we don't need to run it
through chroot - especially when building under a fakeroot style environment.

It already supports rebuilding a directory without having to chroot into it
using the standard "alternate base" -b argument.

The TCL guest busybox is a 32-bit, dynamically linked executable.  This causes
some problems when built on a 64-bit box that doesn't have the 32-bit glibc
installed - typically, a failure to execute with `No such file or directory`.

This uses the guest libraries to execute the process, rather than sticking
with the parent.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>