Ensure the `opt` build directory exists
Somewhere in the last set of changes it became possible to create the initial
bundle file without an `opt` directory under the new build-root overlay.

This led to the replacement `bootsync.sh` script being written as a *file* to
that location, which was then silently replaced with a directory later in the
ISO rebuild process, leaving the original, unmodified `bootsync.sh` process.

This change manually creates the `opt` directory, to ensure that we end up
with the updated build script in place.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>