Put the `gem.list` file into the build root overlay
The final Microkernel is dependent on the `gem.list` file used during gem
mirroring ending up in the final mirror, but that was lost and not noticed
during earlier changes.

This restores the copy into that target location, ensuring that the gem files
are correctly installed by the Microkernel controller on boot.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>