Silence the bundle build process
Now that we are automating the build process, having verbose and human-focused
progress messages from downloads and tar creating result in a much more
difficult to debug build - the errors get drowned in a sea of progress that
nobody watches.

This silences the various tools used to rebuild the ISO image.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>