Shell interpolation is hard, let's use a function!
Trying to get shell interpolation of a multi-word command to work when faced
with some parts of that having spaces is madness.  We did that with busybox
execution using the guest tools, though, since it worked OK in simple cases.

This replaces that with a function that encapsulates the same behaviour in a
much saner fashion.  Now you can run the guest busybox from a 64-bit host
while in a working directory that contains spaces.

Tomorrow the world!

Signed-off-by: Daniel Pittman <daniel@rimspace.net>