Push the working version of the license code.
The great thing about git is that it is happy to push commits from a dirty
working directory.  The bad thing about git is that willingness means I can
push up the older, broken version of a script from before I actually tested
that it worked and not notice.

This moves the copy of the COPYING and LICENSE files to the appropriate part
of the build script, resulting in an end-to-end successful build, rather than
the nasty failure we got with the merged version.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>