Inform users and abort when we can't unpack the ISO image
When running under fakeroot(1), we would try and loopback mount the ISO image
to unpack it despite knowing that could never work. That was a poor user
experience, and we can absolutely do better than that.

This updates the script to check if we are running under fakeroot and skip the
attempt to use mount in favour of informing the user about what they should do
to successfully build.

This closes issue #54.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>