Removes MCollective from the Razor Micro Kernel
MCollective was running as an external service. It used a
static configuration accessible to anyone who accessed
the source, and which could connect to an external master.
This was considered a security threat in the security
evaluation.

This commit removes MCollective completely. The impact of
doing this should be zero since it wasn't currently used.