Limits the services on port 2156 and 2157 to localhost.
The Microkernel exposes two services on ports 2156 and 2157
as part of Tiny Core Linux. The service available on port 2156
handles configuration of the running Microkernel image. It is
unauthenticated, and allows users to load kernel modules and add
new extensions to TCL. The service on 2157 provides information
about the running system.

This commit limits those services to localhost, thus preventing
anyone from accessing them remotely.