Allow reading gem list from a file during mirroring
This extends the gem mirroring script to support reading from a file, as well
as the command line, to find out which set of gems to mirror.

This allows an extended input format in the file, where we specify version
constraints, which resolves #59, a need to specify version constraints.
Now we have a clean mechanism for specifying them.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>