The build process expects a default password, but does not bother to set it
The build process expects a fixed default password in the development and
debug builds, but does not bother to set that. This leads to users, and our CI
system, unexpectedly building ISO images that don't allow login as expected.

Since this is a fixed, documented, default password there is no more security
exposure to be putting it in as the default in the scripts than it does to
manually enter it on every automated, central build.

Ultimately this should probably be eliminated entirely, because it has zero
security value: any attacker can simply read our documentation, and any
automated tool can add whatever password we pick, but for now this improves
utility without surprising our existing users.

This closes #63.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>