The stomp gem was only used to support MCollective
We only included the stomp gem in the system to support MCollective, which is
now gone from the Microkernel.  This removes the gem as well, for a marginal
reduction in size, and less things to care about when figuring out what works
or not in the MK image.

This closes #65.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>