Remove the dangerous "reuse previous download" option
The script that collected data together for the Microkernel image had a "reuse
previous download" option that triggered a particularly dangerous set of
behaviours:

When enabled it assumed that it should simply use the existing build directory
as-is, and should avoid fetching content that already existed in
that directory.

This seems innocuous enough - an optimization you might choose to use during
development - but actually masks a dangerous failure mode.  If we eliminate
something from the build, like we did the stomp gem, it may still be built
into the final image when reusing the downloaded content, because it happens
to sit in the build directory.

A correct version of this would use a cache that sits aside from the build
directory, and only put in place files that would actually be used.

That is more complex to develop, and doesn't really add value compared to
using a simple squid proxy server to cache the downloaded files, so I have
simply expunged the reuse facility.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>