Support multiple .deb package source URLs
The official Debian packages are already broken up into "main", "contrib", and
"non-free" groups, which means that there is no single source of even
'official' packages. This patch adds support for using multiple package lists
in place of a single one.

Note: the order in which lists are specified DOES MATTER. If the same package
is available in multiple repositories, the repository listed first will take
precedence.