Changes download location for MK dependencies
This commit changes the download location for the Hanlon Microkernel
dependencies so that they are downloaded from the GitHub project
itself rather than from my personal Dropbox account (where they've
resided during development) or from the Puppet Labs download
page (which is where they were downloaded when this project was
known as the Razor-Microkernel). Specifically, the files downloaded
from the project "releases" area are as follows:

* mk-open-vm-tools.tar.gz: this file contains a version of the
open-vm-tools utilities compiled and packaged for use with Tiny Core
Linux; it used to be downloaded from the Puppet Labs download site
in the Razor Microkernel days.
* mk-custom-busybox.tar.gz: this file contains a custom version of
the busybox kernel whose DHCP client has been customized to support
three additional DHCP options; specifically:
    * hanlon_server (option 224), which provides a mechanism for
defining an IP address for the Hanlon server that is different
from the next-server IP address (which must point at the server that
provides the bootstrap kernel for the PXE boot process, typically
a TFTP server)
    * hanlon_port (option 225), which provides a mechanism for passing
the port that Hanlon is listening on through the DHCP response sent
back to the Microkernel (important if the Hanlon server is listening
on a port other than the default port, 8026)
    * hanlon_base_uri (option 226), which provides a mechanism for
passing the base URI (or context) under which the Hanlon RESTful
API is accessible (important if Hanlon is deployed to a server
as a WAR file and this context changes from the default /hanlon/api
context defined in Hanlon and the Hanlon-Microkernel)
* mk-ipmi-mods-and-tools.tar.gz: this file contains the drivers
and utilities used by the Microkernel to access the BMC of an
underlying node (if one exists on the node) and report back
the associated IPMI-related facts to Hanlon

with this change in download location for these three files, the
Hanlon-Microkernel should be ready for a 1.0 release